cms栏目管理

+++ jbone-cms/jbone-cms-core/src/main/java/cn/jbone/cms/core/dao/entity/BaseEntity.java
@@ -39,5 +39,5 @@ public class BaseEntity implements Serializable {
      * 创建者
      */
     @Column(name = "creator")
-    private Long creator;
+    private Integer creator;
 }
